+++ src/api/trades.ts
@@ -1,16 +1,9 @@
 import axios from 'axios';
 
-import {Trade} from 'types';
+import {GetTradesParams, Trade} from 'types';
 import {authorizationHeaders} from 'utils/authentication';
 
 const BASE_URL = `${process.env.REACT_APP_API_URL}/api/trades`;
-
-export interface GetTradesParams {
-  buy_order?: number;
-  buy_order__primary_currency__id?: number;
-  buy_order__secondary_currency__id?: number;
-  sell_order?: number;
-}
 
 export const getTrades = async (params?: GetTradesParams): Promise<Trade[]> => {
   try {
