+++ src/dispatchers/followings.ts
@@ -1,7 +1,7 @@
-import {deleteFollower as _deleteFollowing, getFollowers as _getFollowings, GetFollowersParams} from 'api/followers';
+import {deleteFollower as _deleteFollowing, getFollowers as _getFollowings} from 'api/followers';
 import {store} from 'store';
 import {resetFollowing as _resetFollowing, setFollowings, startLoading, unsetFollowing} from 'store/followings';
-import {AppDispatch} from 'types';
+import {AppDispatch, GetFollowersParams} from 'types';
 import {getNextUrlFromState} from 'utils/urls';
 
 export const deleteFollowing = (followerId: number, followingUserId: number) => async (dispatch: AppDispatch) => {
