+++ src/types/api/mints.ts
@@ -1,0 +1,10 @@
+export interface CreateMintRequest {
+  amount: number;
+  currency: number;
+}
+
+export interface GetMintsParams {
+  currency: number;
+  page?: number;
+  page_size?: number;
+}
