--- src/types/api/notifications.ts
@@ -1,3 +1,0 @@
-export interface UpdateNotificationRequest {
-  is_read: boolean;
-}
