+++ src/types/api/trades.ts
@@ -1,0 +1,6 @@
+export interface GetTradesParams {
+  buy_order?: number;
+  buy_order__primary_currency__id?: number;
+  buy_order__secondary_currency__id?: number;
+  sell_order?: number;
+}
